requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
neo4j==5.14.0
requests==2.31.0
beautifulsoup4==4.12.2
pypdf2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8